# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaravelVueErp API",
        "description": "API Documentation for LaravelVueErp Project Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/clientservice/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get all clients with pagination",
                "operationId": "ba6c8bd284f0c63a61b4572a2a6e8c51",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clients retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Client"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "operationId": "8fa346cbeb8167141ce63b5a6c1bfe72",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "phone",
                                    "address",
                                    "city",
                                    "county",
                                    "customer_type",
                                    "lead_source",
                                    "preferred_contact",
                                    "registration_date"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "John Smith"
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "Jane Smith",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.smith@company.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+254712345678"
                                    },
                                    "alt_contact": {
                                        "type": "string",
                                        "example": "+254798765432",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main Street"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Nairobi"
                                    },
                                    "county": {
                                        "type": "string",
                                        "example": "Nairobi County"
                                    },
                                    "postal_address": {
                                        "type": "string",
                                        "example": "P.O. Box 12345",
                                        "nullable": true
                                    },
                                    "customer_type": {
                                        "type": "string",
                                        "enum": [
                                            "individual",
                                            "company",
                                            "organization"
                                        ],
                                        "example": "company"
                                    },
                                    "lead_source": {
                                        "type": "string",
                                        "example": "Website"
                                    },
                                    "preferred_contact": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "phone",
                                            "sms"
                                        ],
                                        "example": "email"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Technology",
                                        "nullable": true
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "Tech Solutions Ltd",
                                        "nullable": true
                                    },
                                    "registration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/clientservice/clients/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Clients"
                ],
                "summary": "Toggle client active status",
                "operationId": "1af73406c355f5e0eb24884aeac74e7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client status toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "type": "string",
                        "example": "John Smith"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "Jane Smith",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.smith@company.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+254712345678"
                    },
                    "alt_contact": {
                        "type": "string",
                        "example": "+254798765432",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main Street"
                    },
                    "city": {
                        "type": "string",
                        "example": "Nairobi"
                    },
                    "county": {
                        "type": "string",
                        "example": "Nairobi County"
                    },
                    "postal_address": {
                        "type": "string",
                        "example": "P.O. Box 12345",
                        "nullable": true
                    },
                    "customer_type": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "company",
                            "organization"
                        ],
                        "example": "company"
                    },
                    "lead_source": {
                        "type": "string",
                        "example": "Website"
                    },
                    "preferred_contact": {
                        "type": "string",
                        "enum": [
                            "email",
                            "phone",
                            "sms"
                        ],
                        "example": "email"
                    },
                    "industry": {
                        "type": "string",
                        "example": "Technology",
                        "nullable": true
                    },
                    "company_name": {
                        "type": "string",
                        "example": "Tech Solutions Ltd",
                        "nullable": true
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}